renamed and reoganized files. Installed PropTypes and  Storybook, fixed contact editor so that it does not force refresh and instead changes contacts array. Fixed formatting issues with the edit function. working on router next
