+++ db.json
@@ -164,11 +164,11 @@
       "id": "d2dcb07d-67b8-4097-b46f-4c366e46371a",
       "first": "Alyssa",
       "last": "Torres",
-      "email": "shortcircuit@spoonflower.com",
-      "dob": "1990-06-19T04:00:00.000Z",
+      "email": "ator@gmail.com",
+      "dob": "1990-06-24T04:00:00.000Z",
       "phone": "9198867885",
       "company": "Spoonflower",
-      "title": "COMRADE IN ARMS AND LEGS"
+      "title": "Associate"
     }
   ]
 }
