+++ package.json
@@ -4,8 +4,10 @@
   "private": true,
   "dependencies": {
     "moment": "^2.22.2",
+    "prop-types": "^15.6.1",
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4",
     "shoelace-css": "^1.0.0-beta24",
     "uuid-v4": "^0.1.0"
@@ -15,9 +17,17 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
-    "start-api": "json-server db.json"
+    "start-api": "json-server db.json",
+    "storybook": "start-storybook -p 9009 -s public",
+    "build-storybook": "build-storybook -s public"
   },
   "devDependencies": {
+    "@storybook/addon-actions": "^3.4.6",
+    "@storybook/addon-links": "^3.4.6",
+    "@storybook/addons": "^3.4.6",
+    "@storybook/react": "^3.4.6",
+    "babel-core": "^6.26.3",
+    "babel-runtime": "^6.26.0",
     "concurrently": "^3.5.1",
     "json-server": "^0.13.0",
     "passport": "^0.4.0",
