+++ src/App.css
@@ -8,6 +8,7 @@
 .main {
   margin: 1rem auto;
   width: 85%;
+  font-family: 'Lato', sans-serif;
   /* font-size: 12px; */
 }
