+++ src/stories/index.js
@@ -1,0 +1,26 @@
+import React from 'react'
+
+import { storiesOf } from '@storybook/react'
+import { action } from '@storybook/addon-actions'
+import { linkTo } from '@storybook/addon-links'
+import 'shoelace-css/dist/shoelace.css'
+
+import { Button, Welcome } from '@storybook/react/demo'
+
+// import Contact from '../Contact'
+import Landing from '../Landing'
+
+storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />)
+
+storiesOf('Button', module)
+  .add('with text', () => <Button onClick={action('clicked')}>Hello Button</Button>)
+  .add('with some emoji', () => (
+    <Button onClick={action('clicked')}>
+      <span role='img' aria-label='so cool'>
+        😀 😎 👍 💯
+      </span>
+    </Button>
+  ))
+
+storiesOf('Landing', module)
+  .add('empty Landing', () => <Landing />)
